requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
uvloop==0.21.0 ; sys_platform != 'win32'